// src/app/(app)/kpi/kpi-22/schemas/kpi22-schemas.ts
import { z } from "zod";

export const kpi22SanitationQuestions = [
  {
    id: "sanitationA",
    label:
      "a) Utility regularly conducts environmental sanitation and hygiene awareness programs and organizes campaigns to promote cleanliness, a healthy environment, and hygiene practices in its service area to protect public health",
    weight: 0.075,
  },
  {
    id: "sanitationB",
    label:
      "b) Utility manages or collaborates with the municipality or other organizations to ensure that sullage generated by households or office buildings is safely reused or disposed of",
    weight: 0.075,
  },
  {
    id: "sanitationC",
    label:
      "c) Utility oversees or works with the municipality or other organizations to ensure that solid waste generated within its service area is disposed of safely",
    weight: 0.075,
  },
  {
    id: "sanitationD",
    label:
      "d) Utility has installed and operates or manages public toilets in its service area, either independently or in collaboration with the municipality or other organizations",
    weight: 0.075,
  },
  {
    id: "sanitationE",
    label:
      "e) Utility oversees the treatment and safe disposal of residential wastewater or fecal sludge produced within its service area, or collaborates with other entities to do so",
    weight: 0.1,
  },
];

export const kpi22LGQuestions = [
  {
    id: "lgA",
    label:
      "a) Utility has conducted a regular meeting with the LG's representative",
    weight: 0.1,
  },
  {
    id: "lgB",
    label: "b) Utility has registered in local government",
    weight: 0.15,
  },
];

export const kpi22RestructuringQuestions = [
  {
    id: "restructuringA",
    label:
      "a) Utility provides drinking water or other water-related services to one or more nearby water utilities, or oversees the entire water supply system",
    weight: 0.1,
  },
  {
    id: "restructuringB",
    label:
      "b) Utility has merged one or more neighboring water utilities into itself to expand the service area and improve water services in the merged areas",
    weight: 0.1,
  },
  {
    id: "restructuringC",
    label:
      "c) Utility manages water supply system and provides water services under the guidance of a Water Board or Water Cooperative, or under some other management model, such as the private sector",
    weight: 0.15,
  },
];

export const kpi22InputSchema = z.object({
  sanitationA: z.enum(["yes", "no"]),
  sanitationB: z.enum(["yes", "no"]),
  sanitationC: z.enum(["yes", "no"]),
  sanitationD: z.enum(["yes", "no"]),
  sanitationE: z.enum(["yes", "no"]),
  lgA: z.enum(["yes", "no"]),
  lgB: z.enum(["yes", "no"]),
  restructuringA: z.enum(["yes", "no"]),
  restructuringB: z.enum(["yes", "no"]),
  restructuringC: z.enum(["yes", "no"]),
});

export type Kpi22Input = z.infer<typeof kpi22InputSchema>;